Setup context for test only once
Assume requirements are not needed for the test.

Requirement
CommandRequirement - is not used now
TableRequirement - most probably can be substituted by
@Before.., @After..
LdapObjectRequirement - to find out how to deal with it.

As time shows, method level requirements are not used in the new code,
and most probably, were implemented for
`Convention based SQL query tests`